Optimizations of models, adding auth manager and uptading menu

+++ Group04_Gym/models/service/FitnessClass.swift
@@ -7,7 +7,7 @@ class FitnessClass: Service {
     var trainerName: String
     var duration: Int
     
-    init(id: String, name: String, totalNumberOfSessions: Int, price: Int, trainerName: String, duration: Int) {
+    init(id: String, name: String, totalNumberOfSessions: Int, price: Double, trainerName: String, duration: Int) {
         self.trainerName = trainerName
         self.duration = duration
         super.init(id: id, name: name, totalNumberOfSessions: totalNumberOfSessions, price: price)
